Retry login when sms() fails
When another computer login to the router, the session is silently
terminated. It is only detected when sending an SMS and getting a status
500.
Changed the code to raise an Error when such scenario happens, this will
make the autologin decorator reconnect.
Since reconnecting in such case could take some time, had to raise the
timeout value to 300 seconds.

Signed-off-by: Amir Vadai <amir@vadai.me>